Bug 1993904 - Acknowledge a possible failure during file copy/move. r=win-reviewers,gstoll
We provide an alternate path if we can't determine whether the file we
are copying or moving is or isn't a directory. This information is only
required to potentially avoid a slow path, so in case we can't get it we
just continue with the copy/move operation and then unconditionally take
the slow path.

Differential Revision: https://phabricator.services.mozilla.com/D268845

UltraBlame original commit: 621bb8c8b8706b532de20e9a2adc217c849f9eae