Bug 1990430 - Make AutoBuffer trivially copyable r=emilio
This was not the case before due to a user-defined destructor. This
fixes the miscompilation by GCC 15.x, but I'm unsure if it was an actual
miscompilation or a consequence of us doing something UB with a
non-trivial type.

Differential Revision: https://phabricator.services.mozilla.com/D268839

UltraBlame original commit: e644100fbd18340f09464cee4f9650dd20e35e19