Bug 1994028 - Introduce ACTION/ACTION_FOR relationship. r=Jamie
This is how we will expose aria-actions internally. It does not map to IA2/ATK relationships, but it is still useful because UIA and MacOS will need to show these actions as a distinct list that is not appended to the standard actions list.

Using relationships also lets us take advantage of a tried-and-true code path for caching and change notification.

Differential Revision: https://phabricator.services.mozilla.com/D268440

UltraBlame original commit: 0c1b45fa8998a63e659a2123a3b913797ddc119d