Bug 1994100 - Add WPTs for mouse/pointer boundary events when `usemap` of `<img>` or `coord` of `<area>` is modified r=smaug,dom-core
It's unclear that how `<area>` is treated within CSS.  However, it's
treated as a transparent object at least when a hit-test of pointing
device events and `<area>` can be a target of mouse events and pointer
events.

When the pointer is moved over the shape of an `<area>`, browsers
dispatch `mouseover` and `pointerover` on the `<area>` instead of the
`<img>`. Then, when it's moved out from the shape, browsers dispatch
`mouseout` and `pointerout`.

However, Chrome and Safari do not dispatch boundary events when the
underneath element of the last pointer position becomes another element.
This does not make sense because modifying the layout, e.g., modifying
the `<img>` size, causes `mouseover` and `pointerover` on the new
element anyway.  Therefore, modifying `usemap` attribute of `<img>`
and `coord` attribute of `<area>` should be treated as a layout change.

Differential Revision: https://phabricator.services.mozilla.com/D268655

UltraBlame original commit: 1ebb5b07c380eb41b87b31739c6da27fb073762f