Bug 1967758 - Add sufficientLength option to DOM text extractor r=translations-reviewers,gregtatum
This patch adds a `sufficientLength` option to the DOM text extractor
logic, which will bail out of text extraction if the extracted text
has surpassed the provided length in code units.

This saves significant resources if the purpose of the extraction does
not need to extract the entirety of the document's text, such as for
Translations language identification.

Differential Revision: https://phabricator.services.mozilla.com/D268284

UltraBlame original commit: ee07c9111ba961e5504cffe460f78b23078379d5