Bug 1994074 - Add Learn more link to safe mode message bars r=rpl
This also corrects the l10n ID for the "learn more" link of message bars
in the extensions panel. Before this patch, the l10n ID of quarantined
domains was also applied to unrelated message bars. Did was not obvious
because the rendered text in either case is "Learn more". However, due
to the use of this l10n ID, a misleading aria-label was set.

Differential Revision: https://phabricator.services.mozilla.com/D268461

UltraBlame original commit: 49bb377a15ea19dd8e5546d815295f04f3deca6d