Bug 1995299 - Remove bad test for names in the Sanitizer. r=smaug
The Sanitizer API used to disallow certain names, but that isn't the
case anymore.

The only reason why the test is currently still passing is accidental:

  const elems_invalid = [
    "", {name: ""},
  ];

"" and {name: ""} are the same empty name, so the constructor throws
because there is a duplicate.

Differential Revision: https://phabricator.services.mozilla.com/D269224

UltraBlame original commit: 94850c926ee57616088fcf7fedd5f62fd067d284