Bug 1959977 - Insert ScriptSource RefPtr to keep it alive until the end of the profiling session r=iain,mstange
We need this API because we need to make sure that the sources of the JS
interpreter, JIT, and the JS execution tracing frames are kept alive until
the serialization.

The implementation is done similarly to `GeckoProfilerRuntime::strings_`.
One difference is that, strings_ is main thread only. But we need to
make sure that the scriptSources_ can be accessed from other threads
since the profile serialization happen in a background thread.

Differential Revision: https://phabricator.services.mozilla.com/D259257

UltraBlame original commit: 9eb467281521ba4414b1a6119e3ace5ea999b414