Bug 1995177. Refactor and improve nsLayoutUtils::ConstrainToCoordValues. r=gfx-reviewers,lsalzman
This patch does a few things at once, it was hard to separate it out into smaller patches.

Combines two functions that are almost the same into one function.

Stops using std::clamp because how it handles NaN is undefined (this can occur).

Keeps the assert that size is non-negative that was only in one version (the next patch in the series handles the one user of this function that can pass in negative).

Allows for NaN with the "size is non-negative" assert (this can occur).

So this function should fully handle all cases with NaN. Added a comment to that effect.

Forces the inputs/outputs to be double. Because nscoord_max can not be represented exactly as a float (so we'd return float(nscoord_max) which we would then convert to nscoord and it would not have the value nscoord_max).

Adapts the only user of this function to the new interface.

Differential Revision: https://phabricator.services.mozilla.com/D269155

UltraBlame original commit: c8e724ba7c14a6436d099b117ecbe25b122a86ca