Bug 1995362 - Part 5: Update enterprise policies to use the new online Suggest pref. r=mkaply,Standard8
Online Firefox Suggest is changing in 146: The old opt-in experience is being
replaced with a new opt-out experience that uses OHTTP to protect user privacy.
As part of that, we are replacing the `quicksuggest.dataCollection.enabled` pref
with a new `quicksuggest.online.enabled` pref. The online checkbox in the
settings UI will now correspond to this new pref.

In case it matters, the old pref defaulted to false (since old online was
opt-in), and the new pref defaults to true (since new online is opt-out).

Depends on D269323

Differential Revision: https://phabricator.services.mozilla.com/D269324

UltraBlame original commit: 64d118c44868fa267b9fe4f041966d4ab5348a82