Bug 1996234 - Package icons in the RPM package. r=releng-reviewers,jcristau DONTBUILD
Without this, icons don't get installed alongside the package and we end
up with the default one for the system instead of the firefox logo, both
on the desktop and as the window icon.

We copy icons instead of using symbolic links because RPM doesn't like
absolute symlinks and emits warnings when we use those. And relative
symlinks are not worth the headache to save a few kilobytes.

Differential Revision: https://phabricator.services.mozilla.com/D269950

UltraBlame original commit: 7cde41eef0616d8b0db5739602eab3cf0aa4fdeb