Bug 1992420 - Do not retry the crash reporter necko backend upon critical failure r=gsvelto
We allow two attempts for the background task to succeed. If it fails
twice (not necessarily in a row), we simply use the curl backend
instead. Some failure modes of the background task are easily observable
(crashes produce signals or specific return values, and the launcher returns
255 on fatal error), however there are other failure modes, like
misconfigured network settings, for which we cannot account. This is why
we use a simpler means of fallback.

This also adds a mocking interface for statics, so that e.g. interior
mutability won't affect concurrently-run tests, and uses it to test the
behavior.

Differential Revision: https://phabricator.services.mozilla.com/D267469

UltraBlame original commit: c6be60553beb693b72ee1f81ca2b50736db59edc