Bug 1994005 - Handle the commit and back buttons for the edit toolbar r=android-reviewers,Roger
When the user presses "Go" in the toolbar we will navigate to the inline
autocomplete suggestion if available or the current query.
Handling the back press in the OS navigation bar is only possible on API33+
which is already supported in the application through Compose BackHandlers.
On lower Android versions the framework does not inform that about the gesture
something that is requested at google in
https://issuetracker.google.com/issues/241705563.

Differential Revision: https://phabricator.services.mozilla.com/D268897

UltraBlame original commit: 69b63b02b0645d80f7cfc311be406bdb75acee13