Bug 1996944 - Remove leftover D2D JS_SetProperty call. r=chutten,toolkit-telemetry-reviewers
The leftover JS_SetProperty for D2D was accidentally setting the D3D11 version property
to a string value, when it was expected to be a number, causing downwind problems. It
needs to be removed.

This adds a test clause to verify the D3D11 version is a number.

Differential Revision: https://phabricator.services.mozilla.com/D270390

UltraBlame original commit: ef4734517e044462b29632c2467e65769ccc1a09