Bug 1996291 - Type check MLEngineChild.sys.mjs r=firefox-ai-ml-reviewers,tarek
This is adding types and type checking to some of our code. I'd like to
do this incrementally as it can really help find bugs and ensure that
our documentation is accurate. The docs for doing this are still in
review, but the linting team is fine with me adding more types to our ml
code as they are working on building out the type inference in Firefox.

The in-review docs are here: https://phabricator.services.mozilla.com/D268571

Differential Revision: https://phabricator.services.mozilla.com/D269980

UltraBlame original commit: 61d730c7b3a8177aa6a4e4c3018630ccfc874549