Bug 1997298 - Use a RW/RX mapping for the remote initialization of sandboxed children. r=bobowen
This removes the need for a remote RWX allocation in the child, or for a
remote change of protection from RW to RX. Instead, sandboxed child
processes only ever get a RX view on the mapping, which the parent
initializes with a RW view.

The calls to WriteProcessMemory do not need to occur anymore, since the
parent updates through the RW view automatically propagate to the
child's RX view.

We also delete ServiceResolverThunk::CopyThunk which is dead code.

Differential Revision: https://phabricator.services.mozilla.com/D270668

UltraBlame original commit: 70602dab118b903c34b53536159d08fa73820cb4