Bug 1997520 - [devtools] Use checkRuleViewContent in browser_rules_at_starting-style.js. r=devtools-reviewers,ochameau.
This allows us to remove the checks that we were doing for overridden properties as those are included in checkRuleViewContent.
checkRuleViewContent is slightly modified: we filter out .registered-properties for now, and we tweak the assertion messages
so failures are easier to debug.

Differential Revision: https://phabricator.services.mozilla.com/D270809

UltraBlame original commit: 34ed759930e38f4bcb13a4ba514b3428143570a9