Bug 1998239 - Fix nsIFrame::GetNearestWidget to not use nsView::GetNearestWidget. r=jwatt
This unifies the popup and non-popup cases in preparation for
bug 1994942.

There's one tricky bit here, which is that we need to fix the
app-units-per-dev-px that we return the coordinates to.

nsView::GetNearestWidget returns the offset in its own coordinates, so
match that. That bug went unnoticed because we don't have zoomed
subdocuments inside popups.

Differential Revision: https://phabricator.services.mozilla.com/D271274

UltraBlame original commit: 7320bf3c5bcfe3a5b126f49c1965f515f240f323