Bug 1991743 - Allow moz-input elements to be used in forms. r=mkennedy
By utilizing the static `formAssociated` property, we allow elements
that extend the MozBaseInputElement to work as expected when used in
form elements. Since we are changing the base input element to be a
form-associated element, we have to implement some additional functions
that are common to form elements. These are the `form()` getter, the
public `setFormValue` method, and the public `formResetCallback` method.

The `setFormValue` is a public method because we need subclasses of the
base input element to be able to modify the private internals of the
base input class.

In order to correctly implement `formResetCallback`, we also added logic
to the MozBaseInputElement to track the default value of the element.
This has the benefit of making elements that extend the base input
behave more like native HTML elements.

Differential Revision: https://phabricator.services.mozilla.com/D268049

UltraBlame original commit: 0cd563373c782ad91cc6d3a3019b6e3762ee46f4