Bug 1968172 - Skip tracing rooted vectors of non-nursery types in minor GC r=jandem
I noticed that for this test case the vector we're tracing repeatedly in
minor GC doesn't ever actually contain any nursery pointers.

The patch adds a base class to the GCPolicy template to add a default
implementation of mightBeInNursery(), returning true. A definition for
GCPolicy<jsid> overrides this since Ids can't be in the nursery. Finally we can
check this in StackGCVector::trace to skip tracing during minor GC for
RootedVectors.

This mostly removes the cost of nursery collection for the test case.

Differential Revision: https://phabricator.services.mozilla.com/D271248

UltraBlame original commit: 1493035f6cab47f2fc738b8e25071b69eaff1915