Bug 1991701 - Re-enable touch on Linux with RFP and remove RFPTarget::PointerId. r=tjr,layout-reviewers,emilio
On Windows, we do not spoof touch-related inputs anymore, and we
pretend touch is always available.
There is no reason not to do the same on Linux, even though the code to
report the number of maxTouchPoints is not implemented there.

Differential Revision: https://phabricator.services.mozilla.com/D267266

UltraBlame original commit: 56c31a3ddb6adc4b9d8dddaa10afaa942ffc6526