Bug 1881130: Add eval-in-global support to setInterruptCallback for use in fuzzing r=jandem
The purpose of this patch is to fix a recurring problem when fuzzing the shell, which is that setInterruptCallback gives power that our actual interrupt handlers don't use. For example: ArrayJoinKernel could iterate over a large array, so we check for interrupts inside the main loop. However, an interrupt callback written in JS can mutate the array we're joining, invalidating ArrayJoinKernel's assumptions. Our real interrupt handlers won't do this. However, we still want some amount of fuzzing coverage for interrupts. Bug 1423937 and bug 1691184 are examples of legitimate bugs that we've found by fuzzing interrupts.

As a compromise, this patch adds support for passing a string as an argument to setInterruptCallback, and evaluating that string in a fresh global. This prevents the interrupt handler from directly mutating the interrupted code, but still allows it to do interesting things (get a backtrace, trigger a GC, etc). When fuzzing, setting a function as an interrupt handler will no longer be supported.

Differential Revision: https://phabricator.services.mozilla.com/D271330

UltraBlame original commit: b63cce36a6b893613eb20d97bb71e7ba324e7c89