Bug 1998419 - Skip using the ScriptPreloader cache when reading in content scripts for newtab. r=home-newtab-reviewers,maxx
This is a safe, upliftable change that gets us off of the ScriptPreloader. The next logical
step is to get rid of the noscript version of the activity-stream HTML page, and load the
scripts on the page "normally" with <script> tags. That will occur instead in bug 1998468.

Differential Revision: https://phabricator.services.mozilla.com/D271425

UltraBlame original commit: 7af8d2e0c24e4c06f6fe9064357a8a8da745828b