Bug 1997896 - Update source zone atom marking bitmaps when gray unmarking regardless of the state of the target zone r=sfink
This fixes a problem with gray unmarking where we return early based on
|zone|'s GC state, potentially skipping updating |sourceZone|'s atom marking
bitmap. The problem is that we still need to do this regardless of the target
zone's state.

I was also able to make a testcase that reproduced gray marking failures which
is fixed by this patch.

Differential Revision: https://phabricator.services.mozilla.com/D271755

UltraBlame original commit: 291116596398ae214fcb0053d602aa220b4ab02d